refactor: 전자지갑 이용 안내 수정 및 컴포넌트 경로 이동
