+++ app/wallet/components/guide/GuideIntroCard.tsx
